+++ CHANGELOG.md
@@ -1,17 +1,21 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/)
 
+## [2.0.16] - 2026-08-18
+
+* refactor from `openid` to `oidc`
+
 ## [2.0.15] - 2026-08-18
 
 * provide optional post-authentication, pre-identity creation user-fn
 
 ## [2.0.14] - 2026-08-18
 
-* provide openid/token-field accessor functions
+* provide oidc/token-field accessor functions
 
 ## [2.0.13] - 2026-08-17
 
-* provide openid/principal helper function
+* provide oidc/principal helper function
 
 ## [2.0.12] - 2026-08-17
 
@@ -38,7 +42,7 @@ All notable changes to this project will be documented in this file. This change
 
 A signficant release that introduces:
 
-* OpenID Authentication
+* OIDC Authentication
   * Authorization Code flow for web-ui
   * Client Credentials flow for api
   * Refresh token redemption